Refactor Changelog link and add to MobileBottomNav
